projeto/requirements.txt
beautifulsoup4==4.12.3
google_api_python_client==2.136.0
google_auth_oauthlib==1.2.0
groq==0.9.0
langchain_community==0.2.6
langchain_huggingface==0.0.3
langchain_text_splitters==0.2.2
protobuf==5.27.2
python-dotenv==1.0.1
Requests==2.32.3
